backend/pyproject.toml
[project]
name = "miz-dac_web"
version = "0.2.1.dev0"
description = "DAC analysis on web, the backend handlers."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "MIZiper"}
]
license = {text = "Apache-2.0"}
license-files = ["LICENSE*"]
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "matplotlib>=3.10.5",
    "miz-dac>=0.4.3",
    "pydantic>=2.11.7",
    "python-jose[cryptography]>=3.3.0",
    "starlette>=0.47.2",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]

[tool.setuptools.package-dir]
dac_web = "dac_web"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dac_web"]
